requirements.txt
streamlit==1.46.0
pandas==2.2.3
yfinance==0.2.63
google-generativeai
python-dotenv
plotly==6.7.0
google-auth